--- frontend/src/components/LoginCard.jsx
@@ -1,154 +1,0 @@
-import { useState } from "react";
-import { useSignIn, SignUpButton } from "@clerk/clerk-react";
-import { useNavigate } from "react-router-dom";
-
-// - On success -> /chat
-export default function Login() {
-  const navigate = useNavigate();
-  const { isLoaded, signIn, setActive } = useSignIn();
-
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState("");
-
-  if (!isLoaded) return null;
-
-  async function onSubmit(e) {
-    e.preventDefault();
-    if (busy) return;
-    setBusy(true);
-    setError("");
-
-    try {
-      const res = await signIn.create({
-        identifier: email,
-        password,
-      });
-
-      if (res.status === "complete") {
-        await setActive({ session: res.createdSessionId });
-        navigate("/chat");
-      } else {
-        // Other steps like 2FA could land here
-        setError("Additional verification required. Please continue the flow.");
-      }
-    } catch (err) {
-      // Collect Clerk error messages nicely
-      const first =
-        err?.errors?.[0]?.longMessage ||
-        err?.errors?.[0]?.message ||
-        "We couldn’t sign you in. Please check your details and try again.";
-      setError(first);
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  async function onForgot(e) {
-    e.preventDefault();
-    setError("");
-    if (!email) {
-      setError("Enter your email above, then press “Forgot?”.");
-      return;
-    }
-    try {
-      await signIn.create({
-        strategy: "reset_password_email_code",
-        identifier: email,
-      });
-      // Clerk will email a code; you can build a /reset route later
-      setError("Password reset email sent. Check your inbox.");
-    } catch (err) {
-      const first =
-        err?.errors?.[0]?.longMessage ||
-        "Could not start password reset. Please try again.";
-      setError(first);
-    }
-  }
-
-  return (
-    <main className="min-h-screen bg-[#fff8f2] flex items-center justify-center p-4">
-      <section
-        className="w-full max-w-md bg-white rounded-2xl shadow-sm ring-1 ring-slate-200 p-6"
-        aria-label="Login form"
-      >
-        {/* top label (mockup shows “placeholder”) */}
-        <p className="text-center text-xl font-semibold mb-2">placeholder</p>
-
-        <h1 className="text-center text-3xl font-bold tracking-tight">
-          Log in
-        </h1>
-        <p className="text-center text-slate-600 mt-1 mb-6">
-          An account is required for AI recognition.
-        </p>
-
-        <form onSubmit={onSubmit} className="space-y-3">
-          <div>
-            <label htmlFor="email" className="sr-only">
-              Email address
-            </label>
-            <input
-              id="email"
-              type="email"
-              autoComplete="email"
-              inputMode="email"
-              placeholder="Email address"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              className="w-full rounded-xl border border-slate-300 px-3 py-3 outline-none focus:ring-2 focus:ring-violet-500"
-              required
-            />
-          </div>
-
-          <div className="relative">
-            <label htmlFor="password" className="sr-only">
-              Password
-            </label>
-            <input
-              id="password"
-              type="password"
-              autoComplete="current-password"
-              placeholder="Password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              className="w-full rounded-xl border border-slate-300 px-3 py-3 pr-20 outline-none focus:ring-2 focus:ring-violet-500"
-              required
-              minLength={8}
-            />
-            <button
-              onClick={onForgot}
-              className="absolute right-3 top-1/2 -translate-y-1/2 text-sm text-slate-600 hover:underline"
-              type="button"
-            >
-              Forgot?
-            </button>
-          </div>
-
-          {error && (
-            <p role="alert" className="text-sm text-red-600">
-              {error}
-            </p>
-          )}
-
-          <button
-            type="submit"
-            disabled={busy}
-            className="w-full rounded-xl bg-black text-white py-3 font-medium disabled:opacity-60"
-          >
-            {busy ? "Signing in…" : "Continue"}
-          </button>
-        </form>
-
-        <p className="text-center text-sm text-slate-700 mt-6">
-          Don’t have an account?{" "}
-          <SignUpButton mode="modal">
-            <span className="font-medium underline cursor-pointer">
-              Sign up
-            </span>
-          </SignUpButton>
-        </p>
-      </section>
-    </main>
-  );
-}
